Extract channel reading from reporter
We now determine the assignment count and list of errors in the
Migrator, and then pass them to GenerateReport instead of passing the
channels directly.

This simplifies GenerateReport and will allow us to have multiple
writers for the error channel, i.e., when the populator is running.

Also:
- Backfilled a test for Migrator
- Extracted channel closing logic from FetchRoleAssignments to
facilitate having multiple writers

[#151616769]

Signed-off-by: Jeff Pak <jeffrey.pak@emc.com>